Fix template method in RPPart and add its test class

+++ app/src/main/java/planning/Text.java
@@ -5,5 +5,5 @@ package planning;
  */
 public interface Text {
 	String getContent() throws Exception;
-	Double getSize();
+	int getSize();
 }
